Vietnamese Telex 1.02 Update
Note: Vietnamese Telex (Legacy Mode) is now deprecated
