+++ legacy/v/vietnamese_telex_legacy/DEPRECATED.md
@@ -1,0 +1,2 @@
+This keyboard has been deprecated and replaced by release/v/vietnamese_telex
+
