+++ legacy/v/vietnamese_telex_legacy/vietnamese_telex_legacy.keyboard_info
@@ -1,0 +1,41 @@
+{
+  "license": "mit",
+  "description": "Telex or TELEX (Vietnamese: Qu\u1ED1c ng\u1EEF \u0111i\u1EC7n t\u00EDn, lit. 'national language telex'), is a convention for encoding Vietnamese text in plain ASCII characters. Originally used for transmitting Vietnamese text over telex systems, it is one of the most used input method on phones and touchscreens and also computers. Vietnamese Morse code uses the TELEX system. In 1993, the use of Telex as an input method was standardized in Vietnam as part of TCVN 5712. This version uses the older convention of tonal diacritic placement for Vietnamese Latin Quoc Ngu alphabet.",
+  "id": "vietnamese_telex_legacy",
+  "languages": {
+    "vi": {
+      "displayName": "Vietnamese",
+      "languageName": "Vietnamese"
+    }
+  },
+  "name": "Vietnamese Telex (Legacy Mode)",
+  "authorName": "Mike Vo",
+  "authorEmail": "info@miketvo.com",
+  "lastModifiedDate": "2022-03-17T01:22:55Z",
+  "sourcePath": "release/v/vietnamese_telex_legacy",
+  "version": "1.0",
+  "packageFilename": "vietnamese_telex_legacy.kmp",
+  "jsFilename": "vietnamese_telex_legacy.js",
+  "encodings": [
+    "unicode"
+  ],
+  "jsFileSize": 4243247,
+  "packageFileSize": 2463925,
+  "packageIncludes": [
+    "visualKeyboard",
+    "welcome",
+    "documentation",
+    "fonts"
+  ],
+  "minKeymanVersion": "10.0",
+  "helpLink": "https://help.keyman.com/keyboard/vietnamese_telex_legacy",
+  "platformSupport": {
+    "windows": "full",
+    "macos": "full",
+    "linux": "full",
+    "android": "full",
+    "ios": "full",
+    "desktopWeb": "full",
+    "mobileWeb": "full"
+  }
+}
