+++ release/v/vietnamese_telex/HISTORY.md
@@ -3,9 +3,16 @@ Vietnamese Telex Change History
 
 1.0 (2022-02-16)
 ----------------
-* Created by Mike Vo
+* Created by Mike Vo.
 
 1.01 (2022-03-16)
 ----------------
-* **Bugfix:** Added missing Telex diacritic rules for gi + a (previously gia + s would incorrectly outputs "gía" instead of "giá"; same with huyền - "f", hỏi "r", ngã "x", and "nặng" j modifiers)
+* **Bugfix:** Added missing Telex diacritic rules for gi + a (previously gia + s would incorrectly outputs "gía" instead of "giá"; same with huyền - "f", hỏi "r", ngã "x", and "nặng" j modifiers).
 
+1.02 (2022-04-16)
+----------------
+* **Feature:** Added smart `ưo` → `ươ` autocorrection when a final consonant character/digraph is present to better emulate the same behavior from [Microsoft's Vietnamese Telex IME](https://support.microsoft.com/vi-vn/windows/cách-cài-đặt-và-sử-dụng-bàn-phím-tiếng-việt-telex-và-number-key-based-968ba67e-0e19-f004-dc5a-abcacef5048b) and the popular third-party [Unikey IME](https://www.unikey.org). This is to help users migrating from those IMEs to get used to this keyboard faster by making use of their existing muscle memory.
+* Added missing `uo-<final consonant(s)>` + `w` → `ươ-<final consonant(s)>` rules. Thanks [u/AdrikIvanov](https://www.reddit.com/user/AdrikIvanov/) from Reddit for letting me know.
+* Added rhymes: `uym`, `uým`, `uỳm`, `uỷm`, `uỹm` and `uỵm` (currently no evidence of usage in either spoken or written Vietnamese, but I included them anyway as to future proof this IME in case those rhymes ever get used).
+* **Feature:** Merged features from my other keyboard [Vietnamese Telex (Legacy Mode)](https://keyman.com/keyboards/vietnamese_telex_legacy). The user can now choose from `Keyman Configuration / Keyboard Layouts / Vietnamese Telex / Keyboard Options` which tonal diacritics placement convention they want to use (See: [Quy tắc đặt dấu thanh trong chữ quốc ngữ](https://vi.wikipedia.org/wiki/Quy_tắc_đặt_dấu_thanh_trong_chữ_quốc_ngữ)). **Note: Because of its popularity, the older convention is now the default setting when the user first install this keyboard.** [Vietnamese Telex (Legacy Mode)](https://keyman.com/keyboards/vietnamese_telex_legacy) is now deprecated.
+* Additional documentation.
