+++ release/v/vietnamese_telex/extras/vn_kmn_compiler/.gitignore
@@ -1,2 +1,3 @@
 /vn_telex/compiled/
 /vn_telex/test/
+/vn_telex_legacy (deprecated)/compiled/
