+++ release/v/vietnamese_telex/extras/vn_kmn_compiler/vn_telex/compile.py
@@ -1,8 +1,10 @@
 from timeit import default_timer as timer
 import vn_telex.utils.progbar as progbar
+from vn_telex.utils.TelexRule import TelexRule
 import vn_telex.utils.charcases as charcases
 import vn_telex.utils.vnrhymes as vnr
-
+import vn_telex.utils.vnrhymes_old as vnro
+import vn_telex.utils.uow_rules as uow
 
 HEADER_PATH = './raw/header.kmn'
 OUT_PATH = './compiled/out.kmn'
@@ -12,7 +14,7 @@ def main():
     start_time = timer()
 
     print('Generating Vietnamese rhymes... ', end='')
-    rhymes = vnr.generate()
+    rhymes = vnro.generate() + vnr.generate() + uow.generate()
     print(f'{len(rhymes)} generated. [DONE]')
 
     print('Generating uppercase and lowercase permutations... ', end='')
@@ -23,13 +25,13 @@ def main():
             base = charcases.apply_case(rhymes[i].base, permutation_line)
             if len(rhymes[i].result) == len(permutation_line):
                 result = charcases.apply_case(rhymes[i].result, permutation_line)
-                rhymes_cases.append(vnr.TelexRhyme(base, rhymes[i].modifier.lower(), result))
-                rhymes_cases.append(vnr.TelexRhyme(base, rhymes[i].modifier.upper(), result))
+                rhymes_cases.append(TelexRule(base, rhymes[i].modifier.lower(), result, kmn_clogic=rhymes[i].kmn_clogic, kmn_ologic=rhymes[i].kmn_ologic))
+                rhymes_cases.append(TelexRule(base, rhymes[i].modifier.upper(), result, kmn_clogic=rhymes[i].kmn_clogic, kmn_ologic=rhymes[i].kmn_ologic))
             elif len(rhymes[i].result) == len(permutation_line) + 1:
                 result = charcases.apply_case(rhymes[i].result[0:-1], permutation_line) + rhymes[i].modifier.lower()
-                rhymes_cases.append(vnr.TelexRhyme(base, rhymes[i].modifier.lower(), result))
+                rhymes_cases.append(TelexRule(base, rhymes[i].modifier.lower(), result, kmn_clogic=rhymes[i].kmn_clogic, kmn_ologic=rhymes[i].kmn_ologic))
                 result = charcases.apply_case(rhymes[i].result[0:-1], permutation_line) + rhymes[i].modifier.upper()
-                rhymes_cases.append(vnr.TelexRhyme(base, rhymes[i].modifier.upper(), result))
+                rhymes_cases.append(TelexRule(base, rhymes[i].modifier.upper(), result, kmn_clogic=rhymes[i].kmn_clogic, kmn_ologic=rhymes[i].kmn_ologic))
             progbar.print_bar(
                 percentage=round(i / len(rhymes) * 100),
                 message=f'({i}/{len(rhymes)}) Processing {rhymes[i].result}'
@@ -39,11 +41,17 @@ def main():
     print('Generating Keyman code... ', end='')
     content = []
     for i in range(len(rhymes_cases)):
-        code = f"    '{rhymes_cases[i].base}' + '{rhymes_cases[i].modifier}' > '{rhymes_cases[i].result}'\n"
+        code = '    '
+        if rhymes_cases[i].kmn_clogic:
+            code += rhymes_cases[i].kmn_clogic + ' '
+        code += f"'{rhymes_cases[i].base}' + '{rhymes_cases[i].modifier}' > '{rhymes_cases[i].result}'"
+        if rhymes_cases[i].kmn_ologic:
+            code += ' ' + rhymes_cases[i].kmn_ologic
+        code += '\n'
         content.append(code)
         progbar.print_bar(
             percentage=round(i / len(rhymes_cases) * 100),
-            message=f'({i}/{len(rhymes_cases)}) {code[4:-1]}'
+            message=f'({i}/{len(rhymes_cases)}) {code.strip()}'
         )
     progbar.print_done(f'Generating Keyman code... {len(content)} line(s) generated.')
