+++ release/v/vietnamese_telex/source/readme.htm
@@ -13,7 +13,7 @@
 </head>
 <body>
 
-<h1>Keyman Vietnamese Telex 1.01</h1>
+<h1>Keyman Vietnamese Telex</h1>
 
 <p>
   Telex or TELEX (Vietnamese: Quốc ngữ điện tín, lit. 'national language telex'), is a convention for encoding Vietnamese text in plain ASCII characters. Originally used for transmitting Vietnamese text over telex systems, it is one of the most used input method on phones and touchscreens and also computers. Vietnamese Morse code uses the TELEX system. In 1993, the use of Telex as an input method was standardized in Vietnam as part of TCVN 5712.
