+++ release/v/vietnamese_telex/source/welcome.htm
@@ -104,7 +104,7 @@
     </tr>
     <tr>
       <td>Ngang (level)</td>
-      <td>z or nothing</td>
+      <td>nothing</td>
       <td>ngang</td>
       <td>ngang</td>
     </tr>
@@ -150,8 +150,8 @@
 </p>
 
 <div class="msg-important">
-  <p><b>Important Notice:</b> This keyboard uses <a href="https://vi.wikipedia.org/wiki/Quy_tắc_đặt_dấu_thanh_trong_chữ_quốc_ngữ#Kiểu_mới">Quy tắc đặt dấu thanh kiểu mới</a>.</p>
-  <p>For <a href="https://vi.wikipedia.org/wiki/Quy_tắc_đặt_dấu_thanh_trong_chữ_quốc_ngữ#Kiểu_cũ">Quy tắc đặt dấu thanh kiểu cũ</a>, use <a href="https://keyman.com/keyboards/vietnamese_telex_legacy">Vietnamese Telex (Legacy Mode)</a>.</p>
+  <p><b>Important Notice:</b> This keyboard uses <a href="https://vi.wikipedia.org/wiki/Quy_tắc_đặt_dấu_thanh_trong_chữ_quốc_ngữ#Kiểu_cũ">Quy tắc đặt dấu thanh kiểu cũ</a> (òa, òe, úy) by default.</p>
+  <p>You can turn on <a href="https://vi.wikipedia.org/wiki/Quy_tắc_đặt_dấu_thanh_trong_chữ_quốc_ngữ#Kiểu_mới">Quy tắc đặt dấu thanh kiểu mới</a> (oà, oè, uý) in <b>Keyman → Configuration → Keyboard Layouts → Vietnamese Telex → Keyboard Options</b></p>
 </div>
 
 <hr>
