+++ release/v/vietnamese_telex/vietnamese_telex.keyboard_info
@@ -3,5 +3,5 @@
     "languages": [
         "vi"
     ],
-	  "description": "Telex or TELEX (Vietnamese: Quốc ngữ điện tín, lit. 'national language telex'), is a convention for encoding Vietnamese text in plain ASCII characters. Originally used for transmitting Vietnamese text over telex systems, it is one of the most used input method on phones and touchscreens and also computers. Vietnamese Morse code uses the TELEX system. In 1993, the use of Telex as an input method was standardized in Vietnam as part of TCVN 5712. This version uses the newer convention of tonal diacritic placement for Vietnamese Latin Quoc Ngu alphabet."
+	  "description": "Telex or TELEX (Vietnamese: Quốc ngữ điện tín, lit. 'national language telex'), is a convention for encoding Vietnamese text in plain ASCII characters. Originally used for transmitting Vietnamese text over telex systems, it is one of the most used input method on phones and touchscreens and also computers. Vietnamese Morse code uses the TELEX system. In 1993, the use of Telex as an input method was standardized in Vietnam as part of TCVN 5712."
 }
