+++ release/v/vietnamese_telex/vietnamese_telex.kpj
@@ -12,7 +12,7 @@
       <ID>id_3b4c3f96658aba183d442ec20392d650</ID>
       <Filename>vietnamese_telex.kmn</Filename>
       <Filepath>source\vietnamese_telex.kmn</Filepath>
-      <FileVersion>1.0</FileVersion>
+      <FileVersion>1.02</FileVersion>
       <FileType>.kmn</FileType>
       <Details>
         <Name>Vietnamese Telex</Name>
@@ -176,6 +176,14 @@
       <Filepath>source\..\..\..\shared\fonts\noto\Latn\NotoSerif-Regular.ttf</Filepath>
       <FileVersion></FileVersion>
       <FileType>.ttf</FileType>
+      <ParentFileID>id_0563ef87d4547ee554bf4abed9b82a67</ParentFileID>
+    </File>
+    <File>
+      <ID>id_4cf401c0bf8ed54b6ec8e5c32e66a318</ID>
+      <Filename>options.htm</Filename>
+      <Filepath>source\options.htm</Filepath>
+      <FileVersion></FileVersion>
+      <FileType>.htm</FileType>
       <ParentFileID>id_0563ef87d4547ee554bf4abed9b82a67</ParentFileID>
     </File>
   </Files>
