--- release/v/vietnamese_telex_legacy/HISTORY.md
@@ -1,6 +1,0 @@
-Vietnamese Telex (Legacy Mode) Change History
-====================
-
-1.0 (2022-03-16)
-----------------
-* Created by Mike Vo
