--- release/v/vietnamese_telex_legacy/README.md
@@ -1,38 +1,0 @@
-Keyman Vietnamese Telex (Legacy Mode) keyboard
-==============
-
-© 2022 Mike Vo
-
-Description
------------
-
-Telex or TELEX (Vietnamese: Quốc ngữ điện tín, lit. 'national language telex'), is a convention for encoding Vietnamese text in plain ASCII characters. Originally used for transmitting Vietnamese text over telex systems, it is one of the most used input method on phones and touchscreens and also computers. Vietnamese Morse code uses the TELEX system. In 1993, the use of Telex as an input method was standardized in Vietnam as part of TCVN 5712.
-
-**Note:** This keyboard uses [Quy tắc đặt dấu thanh kiểu cũ](https://vi.wikipedia.org/wiki/Quy_tắc_đặt_dấu_thanh_trong_chữ_quốc_ngữ#Kiểu_cũ)
-
-For [Quy tắc đặt dấu thanh kiểu mới](https://vi.wikipedia.org/wiki/Quy_tắc_đặt_dấu_thanh_trong_chữ_quốc_ngữ#Kiểu_mới), see [this version](https://github.com/miketvo/keyboards/tree/vietnamese-telex).
-
-Report a bug:
------------
-
-* Via SIL Language Software Community (account required): [u/miketvo](https://community.software.sil.org/u/miketvo/summary)
-* Via email: [info@miketvo.com](mailto:info@miketvo.com)
-
-Links
------
-https://miketvo.com
-
-Supported Platforms
--------------------
- * Windows
- * macOS
- * Linux
- * Web
- * iPhone
- * iPad
- * Android phone
- * Android tablet
- * Mobile devices
- * Desktop devices
- * Tablet devices
-
