--- release/v/vietnamese_telex_legacy/extras/vn_kmn_compiler/.gitignore
@@ -1,3 +1,0 @@
-/vn_telex/compiled/
-/vn_telex/test/
-/vn_telex_legacy/compiled/
