--- release/v/vietnamese_telex_legacy/extras/vn_kmn_compiler/README.md
@@ -1,5 +1,0 @@
-# vn_kmn_compiler
-Keyman keyboard .kmn generator for Vietnamese keyboards
-
-Data sources:
-- List of all vietnamese syllables (content of latin-quoc-ngu-syllables.txt): [All syllables in Vietnamese language, Luong Hieu Thi, 21 Mar 2017](https://www.hieuthi.com/blog/2017/03/21/all-vietnamese-syllables.html)
