--- release/v/vietnamese_telex_legacy/extras/vn_kmn_compiler/scripts/rhymes_filter_regex.txt
@@ -1,1 +1,0 @@
-(b|d|h|l|m|n|p|r|s|t|v|x|đ|tr|th|ch|ph|nh|kh|gi|qu|ngh|ng|gh|g|k|c)(?=[aeiouăâêôơưyáéíóúắấếốớứýàèìòùằầềồờừỳảẻỉỏủẳẩểổởửỷãẽĩõũẵẫễỗỡữỹạẹịọụặậệộợựỵ])
