--- release/v/vietnamese_telex_legacy/extras/vn_kmn_compiler/vn_telex/raw/latin-quoc-ngu-rhymes.txt
@@ -1,861 +1,0 @@
-a
-ac
-ach
-ai
-am
-an
-ang
-anh
-ao
-ap
-at
-au
-ay
-e
-ec
-em
-en
-eng
-eo
-ep
-et
-i
-ia
-ich
-im
-in
-inh
-ip
-it
-iu
-iêc
-iêm
-iên
-iêng
-iêp
-iêt
-iêu
-iếc
-iếm
-iến
-iếng
-iếp
-iết
-iếu
-iềm
-iền
-iềng
-iều
-iểm
-iển
-iểng
-iểu
-iễm
-iễn
-iễng
-iễu
-iệc
-iệm
-iện
-iệng
-iệp
-iệt
-iệu
-o
-oa
-oac
-oach
-oai
-oam
-oan
-oang
-oanh
-oao
-oap
-oat
-oay
-oc
-oe
-oec
-oem
-oen
-oeng
-oeo
-oet
-oi
-om
-on
-ong
-ooc
-oong
-op
-ot
-oà
-oài
-oàm
-oàn
-oàng
-oành
-oào
-oày
-oá
-oác
-oách
-oái
-oám
-oán
-oáng
-oánh
-oáo
-oáp
-oát
-oáy
-oã
-oãi
-oãm
-oãn
-oãng
-oãnh
-oão
-oãy
-oè
-oèm
-oèn
-oèo
-oé
-oéc
-oém
-oén
-oéo
-oét
-oòng
-oóc
-oóng
-oõng
-oăc
-oăm
-oăn
-oăng
-oăt
-oạ
-oạc
-oạch
-oại
-oạm
-oạn
-oạng
-oạnh
-oạo
-oạp
-oạt
-oạy
-oả
-oải
-oảm
-oản
-oảng
-oảnh
-oảo
-oảy
-oắc
-oắm
-oắn
-oắng
-oắt
-oằm
-oằn
-oằng
-oẳm
-oẳn
-oẳng
-oẵm
-oẵn
-oẵng
-oặc
-oặm
-oặn
-oặng
-oặt
-oẹ
-oẹm
-oẹn
-oẹo
-oẹt
-oẻ
-oẻm
-oẻn
-oẻng
-oẻo
-oẽ
-oẽm
-oẽn
-oẽo
-oọc
-oọng
-oỏng
-u
-ua
-uc
-ui
-um
-un
-ung
-up
-ut
-uy
-uya
-uych
-uyn
-uynh
-uyp
-uyt
-uyu
-uyêc
-uyên
-uyêt
-uyếc
-uyến
-uyết
-uyền
-uyển
-uyễn
-uyện
-uyệt
-uân
-uâng
-uât
-uây
-uê
-uêch
-uênh
-uôc
-uôi
-uôm
-uôn
-uông
-uôp
-uôt
-uý
-uýa
-uých
-uýn
-uýnh
-uýp
-uýt
-uýu
-uơ
-uơng
-uấn
-uấng
-uất
-uấy
-uần
-uầng
-uầy
-uẩn
-uẩng
-uẩy
-uẫn
-uẫng
-uẫy
-uận
-uậng
-uật
-uậy
-uắng
-uế
-uếch
-uếnh
-uề
-uềnh
-uều
-uể
-uểnh
-uễ
-uễnh
-uệ
-uệch
-uệnh
-uốc
-uối
-uốm
-uốn
-uống
-uốp
-uốt
-uồi
-uồm
-uồn
-uồng
-uổi
-uổm
-uổn
-uổng
-uỗi
-uỗm
-uỗn
-uỗng
-uộc
-uội
-uộm
-uộn
-uộng
-uộp
-uột
-uớ
-uờ
-uờng
-uở
-uỡ
-uợ
-uỳ
-uỳa
-uỳn
-uỳnh
-uỳu
-uỵ
-uỵa
-uỵch
-uỵn
-uỵnh
-uỵp
-uỵt
-uỵu
-uỷ
-uỷa
-uỷn
-uỷnh
-uỷu
-uỹ
-uỹa
-uỹn
-uỹnh
-uỹu
-y
-ya
-ych
-yn
-ynh
-yp
-yt
-yu
-yêm
-yên
-yêng
-yêt
-yêu
-yếm
-yến
-yếng
-yết
-yếu
-yềm
-yền
-yềng
-yều
-yểm
-yển
-yểng
-yểu
-yễm
-yễn
-yễng
-yễu
-yệm
-yện
-yệng
-yệt
-yệu
-à
-ài
-àm
-àn
-àng
-ành
-ào
-àu
-ày
-á
-ác
-ách
-ái
-ám
-án
-áng
-ánh
-áo
-áp
-át
-áu
-áy
-âc
-âm
-ân
-âng
-âp
-ât
-âu
-ây
-ã
-ãi
-ãm
-ãn
-ãng
-ãnh
-ão
-ãu
-ãy
-è
-èm
-èn
-èng
-èo
-é
-éc
-ém
-én
-éng
-éo
-ép
-ét
-ê
-êc
-êch
-êm
-ên
-êng
-ênh
-êp
-êt
-êu
-ì
-ìa
-ìm
-ìn
-ình
-ìu
-í
-ía
-ích
-ím
-ín
-ính
-íp
-ít
-íu
-ò
-òi
-òm
-òn
-òng
-ó
-óc
-ói
-óm
-ón
-óng
-óp
-ót
-ô
-ôc
-ôi
-ôm
-ôn
-ông
-ôp
-ôt
-õ
-õi
-õm
-õn
-õng
-ù
-ùa
-ùi
-ùm
-ùn
-ùng
-ùy
-ú
-úa
-úc
-úi
-úm
-ún
-úng
-úp
-út
-úy
-ý
-ýa
-ých
-ýn
-ýnh
-ýp
-ýt
-ýu
-ăc
-ăm
-ăn
-ăng
-ăp
-ăt
-ĩ
-ĩa
-ĩm
-ĩn
-ĩnh
-ĩu
-ũ
-ũa
-ũi
-ũm
-ũn
-ũng
-ũy
-ơ
-ơi
-ơm
-ơn
-ơp
-ơt
-ư
-ưa
-ưc
-ưi
-ưm
-ưn
-ưng
-ưt
-ưu
-ươc
-ươi
-ươm
-ươn
-ương
-ươp
-ươt
-ươu
-ước
-ưới
-ướm
-ướn
-ướng
-ướp
-ướt
-ướu
-ười
-ườm
-ườn
-ường
-ườu
-ưởi
-ưởm
-ưởn
-ưởng
-ưởu
-ưỡi
-ưỡm
-ưỡn
-ưỡng
-ưỡu
-ược
-ượi
-ượm
-ượn
-ượng
-ượp
-ượt
-ượu
-ạ
-ạc
-ạch
-ại
-ạm
-ạn
-ạng
-ạnh
-ạo
-ạp
-ạt
-ạu
-ạy
-ả
-ải
-ảm
-ản
-ảng
-ảnh
-ảo
-ảu
-ảy
-ấc
-ấm
-ấn
-ấng
-ấp
-ất
-ấu
-ấy
-ầm
-ần
-ầng
-ầu
-ầy
-ẩm
-ẩn
-ẩng
-ẩu
-ẩy
-ẫm
-ẫn
-ẫng
-ẫu
-ẫy
-ậc
-ậm
-ận
-ậng
-ập
-ật
-ậu
-ậy
-ắc
-ắm
-ắn
-ắng
-ắp
-ắt
-ằm
-ằn
-ằng
-ẳm
-ẳn
-ẳng
-ẵm
-ẵn
-ẵng
-ặc
-ặm
-ặn
-ặng
-ặp
-ặt
-ẹ
-ẹc
-ẹm
-ẹn
-ẹng
-ẹo
-ẹp
-ẹt
-ẻ
-ẻm
-ẻn
-ẻng
-ẻo
-ẽ
-ẽm
-ẽn
-ẽng
-ẽo
-ế
-ếc
-ếch
-ếm
-ến
-ếng
-ếnh
-ếp
-ết
-ếu
-ề
-ềm
-ền
-ềng
-ềnh
-ều
-ể
-ểm
-ển
-ểng
-ểnh
-ểu
-ễ
-ễm
-ễn
-ễng
-ễnh
-ễu
-ệ
-ệc
-ệch
-ệm
-ện
-ệng
-ệnh
-ệp
-ệt
-ệu
-ỉ
-ỉa
-ỉm
-ỉn
-ỉnh
-ỉu
-ị
-ịa
-ịch
-ịm
-ịn
-ịnh
-ịp
-ịt
-ịu
-ọ
-ọc
-ọi
-ọm
-ọn
-ọng
-ọp
-ọt
-ỏ
-ỏi
-ỏm
-ỏn
-ỏng
-ố
-ốc
-ối
-ốm
-ốn
-ống
-ốp
-ốt
-ồ
-ồi
-ồm
-ồn
-ồng
-ổ
-ổi
-ổm
-ổn
-ổng
-ỗ
-ỗi
-ỗm
-ỗn
-ỗng
-ộ
-ộc
-ội
-ộm
-ộn
-ộng
-ộp
-ột
-ớ
-ới
-ớm
-ớn
-ớp
-ớt
-ờ
-ời
-ờm
-ờn
-ở
-ởi
-ởm
-ởn
-ỡ
-ỡi
-ỡm
-ỡn
-ợ
-ợi
-ợm
-ợn
-ợp
-ợt
-ụ
-ụa
-ục
-ụi
-ụm
-ụn
-ụng
-ụp
-ụt
-ụy
-ủ
-ủa
-ủi
-ủm
-ủn
-ủng
-ủy
-ứ
-ứa
-ức
-ứi
-ứm
-ứn
-ứng
-ứt
-ứu
-ừ
-ừa
-ừi
-ừm
-ừn
-ừng
-ừu
-ử
-ửa
-ửi
-ửm
-ửn
-ửng
-ửu
-ữ
-ữa
-ữi
-ữm
-ữn
-ững
-ữu
-ự
-ựa
-ực
-ựi
-ựm
-ựn
-ựng
-ựt
-ựu
-ỳ
-ỳa
-ỳn
-ỳnh
-ỳu
-ỵ
-ỵa
-ỵch
-ỵn
-ỵnh
-ỵp
-ỵt
-ỵu
-ỷ
-ỷa
-ỷn
-ỷnh
-ỷu
-ỹ
-ỹa
-ỹn
-ỹnh
-ỹu
