--- release/v/vietnamese_telex_legacy/source/readme.htm
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <title>Vietnamese Telex (Legacy Mode)</title>
-  <style type="text/css">
-    p { font: 10pt Tahoma; }
-    h1 { font: bold 16pt Tahoma; color: #4444cc; margin-bottom: 2px }
-    h2 { font: bold 12pt Tahoma; color: #4444cc; }
-    .msg-important { color: rgb(238, 0, 0); background-color: rgb(255, 213, 213); border: 2px dashed rgb(238, 0, 0); padding: 8pt 16pt; margin: 8pt 0; display: inline-block; }
-  </style>
-</head>
-<body>
-
-<h1>Keyman Vietnamese Telex (Legacy Mode)</h1>
-
-<p>
-  Telex or TELEX (Vietnamese: Quốc ngữ điện tín, lit. 'national language telex'), is a convention for encoding Vietnamese text in plain ASCII characters. Originally used for transmitting Vietnamese text over telex systems, it is one of the most used input method on phones and touchscreens and also computers. Vietnamese Morse code uses the TELEX system. In 1993, the use of Telex as an input method was standardized in Vietnam as part of TCVN 5712.
-</p>
-
-<div class="msg-important">
-  <p><b>Important Notice:</b> This keyboard uses <a href="https://vi.wikipedia.org/wiki/Quy_tắc_đặt_dấu_thanh_trong_chữ_quốc_ngữ#Kiểu_cũ">Quy tắc đặt dấu thanh kiểu cũ</a>.</p>
-  <p>For <a href="https://vi.wikipedia.org/wiki/Quy_tắc_đặt_dấu_thanh_trong_chữ_quốc_ngữ#Kiểu_mới">Quy tắc đặt dấu thanh kiểu mới</a>, use <a href="https://keyman.com/keyboards/vietnamese_telex">this version</a>.</p>
-</div>
-
-<hr>
-
-<h3>Report a bug:</h3>
-<ul>
-  <li>Via SIL Language Software Community (account required): <a href="https://community.software.sil.org/u/miketvo/summary">u/miketvo</a></li>
-  <li>Via email: <a href="mailto:info@miketvo.com">info@miketvo.com</a></li>
-</ul>
-
-<p><a href="https://github.com/keymanapp/keyboards/blob/master/release/v/vietnamese_telex_legacy/HISTORY.md">Version history</a></p>
-
-<hr>
-
-<p>© 2022 Mike Vo</p>
-
-</body>
-</html>
