--- release/v/vietnamese_telex_legacy/source/vietnamese_telex_legacy.kps
@@ -1,140 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Package>
-  <System>
-    <KeymanDeveloperVersion>14.0.286.0</KeymanDeveloperVersion>
-    <FileVersion>7.0</FileVersion>
-  </System>
-  <Options>
-    <ExecuteProgram></ExecuteProgram>
-    <ReadMeFile>readme.htm</ReadMeFile>
-    <GraphicFile>splash.bmp</GraphicFile>
-    <MSIFileName></MSIFileName>
-    <MSIOptions></MSIOptions>
-    <FollowKeyboardVersion/>
-  </Options>
-  <StartMenu>
-    <Folder></Folder>
-    <AddUninstallEntry/>
-    <Items>
-      <Item>
-        <Name>(new)</Name>
-        <FileName>(Product Configuration)</FileName>
-        <Arguments></Arguments>
-        <Icon></Icon>
-        <Location>psmelStartMenu</Location>
-      </Item>
-    </Items>
-  </StartMenu>
-  <Info>
-    <Name URL="">Vietnamese Telex (Legacy Mode)</Name>
-    <Copyright URL="">© 2022 Mike Vo</Copyright>
-    <Author URL="mailto:info@miketvo.com">Mike Vo</Author>
-    <Version URL=""></Version>
-    <WebSite URL="https://miketvo.com">https://miketvo.com</WebSite>
-  </Info>
-  <Files>
-    <File>
-      <Name>..\build\vietnamese_telex_legacy.kmx</Name>
-      <Description></Description>
-      <CopyLocation>0</CopyLocation>
-      <FileType>.kmx</FileType>
-    </File>
-    <File>
-      <Name>..\build\vietnamese_telex_legacy.js</Name>
-      <Description></Description>
-      <CopyLocation>0</CopyLocation>
-      <FileType>.js</FileType>
-    </File>
-    <File>
-      <Name>..\build\vietnamese_telex_legacy.kvk</Name>
-      <Description></Description>
-      <CopyLocation>0</CopyLocation>
-      <FileType>.kvk</FileType>
-    </File>
-    <File>
-      <Name>welcome.htm</Name>
-      <Description></Description>
-      <CopyLocation>0</CopyLocation>
-      <FileType>.htm</FileType>
-    </File>
-    <File>
-      <Name>readme.htm</Name>
-      <Description></Description>
-      <CopyLocation>0</CopyLocation>
-      <FileType>.htm</FileType>
-    </File>
-    <File>
-      <Name>splash.bmp</Name>
-      <Description>File splash.bmp</Description>
-      <CopyLocation>0</CopyLocation>
-      <FileType>.bmp</FileType>
-    </File>
-    <File>
-      <Name>vietnamese_telex_legacy.ico</Name>
-      <Description>File vietnamese_telex_legacy.ico</Description>
-      <CopyLocation>0</CopyLocation>
-      <FileType>.ico</FileType>
-    </File>
-    <File>
-      <Name>..\..\..\shared\fonts\noto\Latn\NotoSans-Bold.ttf</Name>
-      <Description>Font Noto Sans Bold</Description>
-      <CopyLocation>0</CopyLocation>
-      <FileType>.ttf</FileType>
-    </File>
-    <File>
-      <Name>..\..\..\shared\fonts\noto\Latn\NotoSans-BoldItalic.ttf</Name>
-      <Description>Font Noto Sans Bold Italic</Description>
-      <CopyLocation>0</CopyLocation>
-      <FileType>.ttf</FileType>
-    </File>
-    <File>
-      <Name>..\..\..\shared\fonts\noto\Latn\NotoSans-Italic.ttf</Name>
-      <Description>Font Noto Sans Italic</Description>
-      <CopyLocation>0</CopyLocation>
-      <FileType>.ttf</FileType>
-    </File>
-    <File>
-      <Name>..\..\..\shared\fonts\noto\Latn\NotoSans-Regular.ttf</Name>
-      <Description>Font Noto Sans</Description>
-      <CopyLocation>0</CopyLocation>
-      <FileType>.ttf</FileType>
-    </File>
-    <File>
-      <Name>..\..\..\shared\fonts\noto\Latn\NotoSerif-Bold.ttf</Name>
-      <Description>Font Noto Serif Bold</Description>
-      <CopyLocation>0</CopyLocation>
-      <FileType>.ttf</FileType>
-    </File>
-    <File>
-      <Name>..\..\..\shared\fonts\noto\Latn\NotoSerif-BoldItalic.ttf</Name>
-      <Description>Font Noto Serif Bold Italic</Description>
-      <CopyLocation>0</CopyLocation>
-      <FileType>.ttf</FileType>
-    </File>
-    <File>
-      <Name>..\..\..\shared\fonts\noto\Latn\NotoSerif-Italic.ttf</Name>
-      <Description>Font Noto Serif Italic</Description>
-      <CopyLocation>0</CopyLocation>
-      <FileType>.ttf</FileType>
-    </File>
-    <File>
-      <Name>..\..\..\shared\fonts\noto\Latn\NotoSerif-Regular.ttf</Name>
-      <Description>Font Noto Serif</Description>
-      <CopyLocation>0</CopyLocation>
-      <FileType>.ttf</FileType>
-    </File>
-  </Files>
-  <Keyboards>
-    <Keyboard>
-      <Name>Vietnamese Telex (Legacy Mode)</Name>
-      <ID>vietnamese_telex_legacy</ID>
-      <OSKFont>..\..\..\shared\fonts\noto\Latn\NotoSans-Regular.ttf</OSKFont>
-      <DisplayFont>..\..\..\shared\fonts\noto\Latn\NotoSans-Regular.ttf</DisplayFont>
-      <Version>1.0</Version>
-      <Languages>
-        <Language ID="vi">Vietnamese</Language>
-      </Languages>
-    </Keyboard>
-  </Keyboards>
-  <Strings/>
-</Package>
