--- release/v/vietnamese_telex_legacy/source/vietnamese_telex_legacy.kvks
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<visualkeyboard>
-  <header>
-    <version>10.0</version>
-    <kbdname>vietnamese_telex_legacy</kbdname>
-    <flags>
-      <displayunderlying/>
-    </flags>
-  </header>
-</visualkeyboard>
