--- release/v/vietnamese_telex_legacy/vietnamese_telex_legacy.keyboard_info
@@ -1,7 +1,0 @@
-{
-    "license": "mit",
-    "languages": [
-        "vi"
-    ],
-	  "description": "Telex or TELEX (Vietnamese: Quốc ngữ điện tín, lit. 'national language telex'), is a convention for encoding Vietnamese text in plain ASCII characters. Originally used for transmitting Vietnamese text over telex systems, it is one of the most used input method on phones and touchscreens and also computers. Vietnamese Morse code uses the TELEX system. In 1993, the use of Telex as an input method was standardized in Vietnam as part of TCVN 5712. This version uses the older convention of tonal diacritic placement for Vietnamese Latin Quoc Ngu alphabet."
-}
