--- release/v/vietnamese_telex_legacy/vietnamese_telex_legacy.kpj
@@ -1,182 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<KeymanDeveloperProject>
-  <Options>
-    <BuildPath>$PROJECTPATH\build</BuildPath>
-    <CompilerWarningsAsErrors>True</CompilerWarningsAsErrors>
-    <WarnDeprecatedCode>True</WarnDeprecatedCode>
-    <CheckFilenameConventions>True</CheckFilenameConventions>
-    <ProjectType>keyboard</ProjectType>
-  </Options>
-  <Files>
-    <File>
-      <ID>id_3b4c3f96658aba183d442ec20392d650</ID>
-      <Filename>vietnamese_telex_legacy.kmn</Filename>
-      <Filepath>source\vietnamese_telex_legacy.kmn</Filepath>
-      <FileVersion>1.0</FileVersion>
-      <FileType>.kmn</FileType>
-      <Details>
-        <Name>Vietnamese Telex (Legacy Mode)</Name>
-        <Copyright>© 2022 Mike Vo</Copyright>
-      </Details>
-    </File>
-    <File>
-      <ID>id_0563ef87d4547ee554bf4abed9b82a67</ID>
-      <Filename>vietnamese_telex_legacy.kps</Filename>
-      <Filepath>source\vietnamese_telex_legacy.kps</Filepath>
-      <FileVersion></FileVersion>
-      <FileType>.kps</FileType>
-      <Details>
-        <Name>Vietnamese Telex (Legacy Mode)</Name>
-        <Copyright>© 2022 Mike Vo</Copyright>
-      </Details>
-    </File>
-    <File>
-      <ID>id_ede98e4633e239f933cbfd1f4e1b766c</ID>
-      <Filename>HISTORY.md</Filename>
-      <Filepath>HISTORY.md</Filepath>
-      <FileVersion></FileVersion>
-      <FileType>.md</FileType>
-    </File>
-    <File>
-      <ID>id_53e892b8b41cc4caece1cfd5ef21d6e7</ID>
-      <Filename>LICENSE.md</Filename>
-      <Filepath>LICENSE.md</Filepath>
-      <FileVersion></FileVersion>
-      <FileType>.md</FileType>
-    </File>
-    <File>
-      <ID>id_0730bb7c2e8f9ea2438b52e419dd86c9</ID>
-      <Filename>README.md</Filename>
-      <Filepath>README.md</Filepath>
-      <FileVersion></FileVersion>
-      <FileType>.md</FileType>
-    </File>
-    <File>
-      <ID>id_dd9199cf7af8b59e4ddfd90854fa5f07</ID>
-      <Filename>vietnamese_telex_legacy.keyboard_info</Filename>
-      <Filepath>vietnamese_telex_legacy.keyboard_info</Filepath>
-      <FileVersion></FileVersion>
-      <FileType>.keyboard_info</FileType>
-    </File>
-    <File>
-      <ID>id_eaaec85863b3bcf4c92564be19a5960f</ID>
-      <Filename>vietnamese_telex_legacy.kmx</Filename>
-      <Filepath>source\..\build\vietnamese_telex_legacy.kmx</Filepath>
-      <FileVersion></FileVersion>
-      <FileType>.kmx</FileType>
-      <ParentFileID>id_0563ef87d4547ee554bf4abed9b82a67</ParentFileID>
-    </File>
-    <File>
-      <ID>id_e163d991c362d4e0c20cfbae512a22b0</ID>
-      <Filename>vietnamese_telex_legacy.js</Filename>
-      <Filepath>source\..\build\vietnamese_telex_legacy.js</Filepath>
-      <FileVersion></FileVersion>
-      <FileType>.js</FileType>
-      <ParentFileID>id_0563ef87d4547ee554bf4abed9b82a67</ParentFileID>
-    </File>
-    <File>
-      <ID>id_ed908375e1dabcf1f0a2829562df9faf</ID>
-      <Filename>vietnamese_telex_legacy.kvk</Filename>
-      <Filepath>source\..\build\vietnamese_telex_legacy.kvk</Filepath>
-      <FileVersion></FileVersion>
-      <FileType>.kvk</FileType>
-      <ParentFileID>id_0563ef87d4547ee554bf4abed9b82a67</ParentFileID>
-    </File>
-    <File>
-      <ID>id_356e5d149c1e539356d72698c1e401a6</ID>
-      <Filename>welcome.htm</Filename>
-      <Filepath>source\welcome.htm</Filepath>
-      <FileVersion></FileVersion>
-      <FileType>.htm</FileType>
-      <ParentFileID>id_0563ef87d4547ee554bf4abed9b82a67</ParentFileID>
-    </File>
-    <File>
-      <ID>id_8da344c4cea6f467013357fe099006f5</ID>
-      <Filename>readme.htm</Filename>
-      <Filepath>source\readme.htm</Filepath>
-      <FileVersion></FileVersion>
-      <FileType>.htm</FileType>
-      <ParentFileID>id_0563ef87d4547ee554bf4abed9b82a67</ParentFileID>
-    </File>
-    <File>
-      <ID>id_48e3c918d804bfd60a0122dff7f72434</ID>
-      <Filename>splash.bmp</Filename>
-      <Filepath>source\splash.bmp</Filepath>
-      <FileVersion></FileVersion>
-      <FileType>.bmp</FileType>
-      <ParentFileID>id_0563ef87d4547ee554bf4abed9b82a67</ParentFileID>
-    </File>
-    <File>
-      <ID>id_5c1003dc05d55963d85504bc98ef2a9f</ID>
-      <Filename>vietnamese_telex_legacy.ico</Filename>
-      <Filepath>source\vietnamese_telex_legacy.ico</Filepath>
-      <FileVersion></FileVersion>
-      <FileType>.ico</FileType>
-      <ParentFileID>id_0563ef87d4547ee554bf4abed9b82a67</ParentFileID>
-    </File>
-    <File>
-      <ID>id_cb228b37297a2270ac89e3e6e32e4988</ID>
-      <Filename>NotoSans-Bold.ttf</Filename>
-      <Filepath>source\..\..\..\shared\fonts\noto\Latn\NotoSans-Bold.ttf</Filepath>
-      <FileVersion></FileVersion>
-      <FileType>.ttf</FileType>
-      <ParentFileID>id_0563ef87d4547ee554bf4abed9b82a67</ParentFileID>
-    </File>
-    <File>
-      <ID>id_ccfeba9f0553ac8619d54663f4dccb48</ID>
-      <Filename>NotoSans-BoldItalic.ttf</Filename>
-      <Filepath>source\..\..\..\shared\fonts\noto\Latn\NotoSans-BoldItalic.ttf</Filepath>
-      <FileVersion></FileVersion>
-      <FileType>.ttf</FileType>
-      <ParentFileID>id_0563ef87d4547ee554bf4abed9b82a67</ParentFileID>
-    </File>
-    <File>
-      <ID>id_320b047fa5373e3d7bcccbccecfdb91e</ID>
-      <Filename>NotoSans-Italic.ttf</Filename>
-      <Filepath>source\..\..\..\shared\fonts\noto\Latn\NotoSans-Italic.ttf</Filepath>
-      <FileVersion></FileVersion>
-      <FileType>.ttf</FileType>
-      <ParentFileID>id_0563ef87d4547ee554bf4abed9b82a67</ParentFileID>
-    </File>
-    <File>
-      <ID>id_8e28fe50b9ee816383fdbfa5e8b4addd</ID>
-      <Filename>NotoSans-Regular.ttf</Filename>
-      <Filepath>source\..\..\..\shared\fonts\noto\Latn\NotoSans-Regular.ttf</Filepath>
-      <FileVersion></FileVersion>
-      <FileType>.ttf</FileType>
-      <ParentFileID>id_0563ef87d4547ee554bf4abed9b82a67</ParentFileID>
-    </File>
-    <File>
-      <ID>id_d19a4050cbacbba3cfd2e689daf4a622</ID>
-      <Filename>NotoSerif-Bold.ttf</Filename>
-      <Filepath>source\..\..\..\shared\fonts\noto\Latn\NotoSerif-Bold.ttf</Filepath>
-      <FileVersion></FileVersion>
-      <FileType>.ttf</FileType>
-      <ParentFileID>id_0563ef87d4547ee554bf4abed9b82a67</ParentFileID>
-    </File>
-    <File>
-      <ID>id_ce73d6a765ca48d5663e644b59e1c2d5</ID>
-      <Filename>NotoSerif-BoldItalic.ttf</Filename>
-      <Filepath>source\..\..\..\shared\fonts\noto\Latn\NotoSerif-BoldItalic.ttf</Filepath>
-      <FileVersion></FileVersion>
-      <FileType>.ttf</FileType>
-      <ParentFileID>id_0563ef87d4547ee554bf4abed9b82a67</ParentFileID>
-    </File>
-    <File>
-      <ID>id_4aebba9409e2af67c310a07e37a6086c</ID>
-      <Filename>NotoSerif-Italic.ttf</Filename>
-      <Filepath>source\..\..\..\shared\fonts\noto\Latn\NotoSerif-Italic.ttf</Filepath>
-      <FileVersion></FileVersion>
-      <FileType>.ttf</FileType>
-      <ParentFileID>id_0563ef87d4547ee554bf4abed9b82a67</ParentFileID>
-    </File>
-    <File>
-      <ID>id_13c16ffcb5785b9c213129b730d641fe</ID>
-      <Filename>NotoSerif-Regular.ttf</Filename>
-      <Filepath>source\..\..\..\shared\fonts\noto\Latn\NotoSerif-Regular.ttf</Filepath>
-      <FileVersion></FileVersion>
-      <FileType>.ttf</FileType>
-      <ParentFileID>id_0563ef87d4547ee554bf4abed9b82a67</ParentFileID>
-    </File>
-  </Files>
-</KeymanDeveloperProject>
